requirements.txt
streamlit
pandas
numpy
yfinance
requests
scipy
statsmodels
arch
plotly
python-dotenv
